Implementation of scrub_params/2 plug.
